Share tests between BaseRecalibratorIntegrationTest and the Spark equivalent.
